CoreInterface: Rename files and classes
