+++ src/core/entry/core_text_displayer_interface.ts
@@ -1,25 +1,25 @@
-import log from "../../../log";
-import type { ITextDisplayerData } from "../../../main_thread/types";
+import log from "../../log";
+import type { ITextDisplayerData } from "../../main_thread/types";
+import type { IRange } from "../../utils/ranges";
+import { CancellationError } from "../../utils/task_canceller";
+import type { ITextDisplayerInterface } from "../segment_sinks";
 import type {
-  IRemoveTextDataWorkerMessage,
-  IStopTextDisplayerWorkerMessage,
-  IResetTextDisplayerWorkerMessage,
-  IPushTextDataWorkerMessage,
-} from "../../../multithread_types";
-import { WorkerMessageType } from "../../../multithread_types";
-import type { IRange } from "../../../utils/ranges";
-import { CancellationError } from "../../../utils/task_canceller";
-import type { ITextDisplayerInterface } from "../../segment_sinks";
+  IPushTextDataCoreMessage,
+  IRemoveTextDataCoreMessage,
+  IResetTextDisplayerCoreMessage,
+  IStopTextDisplayerCoreMessage,
+} from "../types";
+import { CoreMessageType } from "../types";
 
 /**
- * Implementation of an `ITextDisplayerInterface` running in a WebWorker
- * (so, in a different thread than the `ITextDisplayer`).
+ * Implementation of an `ITextDisplayerInterface` for the Core, that can run in
+ * a WebWorker (so, in a different thread than the `ITextDisplayer`).
  *
- * @class WorkerTextDisplayerInterface
+ * @class CoreTextDisplayerInterface
  */
-export default class WorkerTextDisplayerInterface implements ITextDisplayerInterface {
+export default class CoreTextDisplayerInterface implements ITextDisplayerInterface {
   private _contentId: string;
-  private _messageSender: (msg: IWorkerTextDisplayerInterfaceMessage) => void;
+  private _messageSender: (msg: ICoreTextDisplayerInterfaceMessage) => void;
   public _queues: {
     pushTextData: Array<{
       resolve: (ranges: IRange[]) => void;
@@ -37,7 +37,7 @@ export default class WorkerTextDisplayerInterface implements ITextDisplayerInter
    */
   constructor(
     contentId: string,
-    messageSender: (msg: IWorkerTextDisplayerInterfaceMessage) => void,
+    messageSender: (msg: ICoreTextDisplayerInterfaceMessage) => void,
   ) {
     this._contentId = contentId;
     this._messageSender = messageSender;
@@ -50,7 +50,7 @@ export default class WorkerTextDisplayerInterface implements ITextDisplayerInter
   public pushTextData(infos: ITextDisplayerData): Promise<IRange[]> {
     return new Promise((resolve, reject) => {
       this._messageSender({
-        type: WorkerMessageType.PushTextData,
+        type: CoreMessageType.PushTextData,
         contentId: this._contentId,
         value: infos,
       });
@@ -64,7 +64,7 @@ export default class WorkerTextDisplayerInterface implements ITextDisplayerInter
   public remove(start: number, end: number): Promise<IRange[]> {
     return new Promise((resolve, reject) => {
       this._messageSender({
-        type: WorkerMessageType.RemoveTextData,
+        type: CoreMessageType.RemoveTextData,
         contentId: this._contentId,
         value: { start, end },
       });
@@ -77,7 +77,7 @@ export default class WorkerTextDisplayerInterface implements ITextDisplayerInter
    */
   public reset(): void {
     this._messageSender({
-      type: WorkerMessageType.ResetTextDisplayer,
+      type: CoreMessageType.ResetTextDisplayer,
       contentId: this._contentId,
       value: null,
     });
@@ -89,7 +89,7 @@ export default class WorkerTextDisplayerInterface implements ITextDisplayerInter
    */
   public stop(): void {
     this._messageSender({
-      type: WorkerMessageType.StopTextDisplayer,
+      type: CoreMessageType.StopTextDisplayer,
       contentId: this._contentId,
       value: null,
     });
@@ -155,8 +155,8 @@ export default class WorkerTextDisplayerInterface implements ITextDisplayerInter
   }
 }
 
-type IWorkerTextDisplayerInterfaceMessage =
-  | IPushTextDataWorkerMessage
-  | IRemoveTextDataWorkerMessage
-  | IStopTextDisplayerWorkerMessage
-  | IResetTextDisplayerWorkerMessage;
+type ICoreTextDisplayerInterfaceMessage =
+  | IPushTextDataCoreMessage
+  | IRemoveTextDataCoreMessage
+  | IStopTextDisplayerCoreMessage
+  | IResetTextDisplayerCoreMessage;
