+++ src/core/entry/index.ts
@@ -1,0 +1,4 @@
+import type { IMessageReceiverCallback } from "./core_entry";
+import initializeCoreEntry from "./core_entry";
+export default initializeCoreEntry;
+export type { IMessageReceiverCallback };
