--- src/core/main/worker/index.ts
@@ -1,4 +1,0 @@
-import type { IMessageReceiverCallback } from "./worker_main";
-import initializeWorkerMain from "./worker_main";
-export default initializeWorkerMain;
-export type { IMessageReceiverCallback };
