+++ src/experimental/features/__tests__/local.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from "vitest";
-import initializeWorkerMain from "../../../core/main/worker";
+import initializeCoreEntry from "../../../core/entry";
 import type { IFeaturesObject } from "../../../features/types";
 import { MonoThreadCoreInterface } from "../../../main_thread/core_interface/monothread";
-import MultiThreadContentInitializer from "../../../main_thread/init/multi_thread_content_initializer";
+import MediaSourceContentInitializer from "../../../main_thread/init/media_source_content_initializer";
 import local from "../../../transports/local";
 import addLocalManifestFeature from "../local";
 
@@ -13,15 +13,15 @@ describe("Features list - LOCAL_MANIFEST", () => {
     expect(featureObject).toEqual({
       transports: { local },
       monothread: {
-        init: MultiThreadContentInitializer,
+        init: MediaSourceContentInitializer,
         coreInterface: MonoThreadCoreInterface,
-        workerMain: initializeWorkerMain,
+        initializeCoreEntry,
       },
     });
     expect(featureObject.transports.local).toBe(local);
     expect(featureObject.monothread).not.toBe(null);
-    expect(featureObject.monothread?.init).toBe(MultiThreadContentInitializer);
+    expect(featureObject.monothread?.init).toBe(MediaSourceContentInitializer);
     expect(featureObject.monothread?.coreInterface).toBe(MonoThreadCoreInterface);
-    expect(featureObject.monothread?.workerMain).toBe(initializeWorkerMain);
+    expect(featureObject.monothread?.initializeCoreEntry).toBe(initializeCoreEntry);
   });
 });
