+++ src/experimental/features/__tests__/metaplaylist.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect } from "vitest";
-import initializeWorkerMain from "../../../core/main/worker";
+import initializeCoreEntry from "../../../core/entry";
 import type { IFeaturesObject } from "../../../features/types";
 import { MonoThreadCoreInterface } from "../../../main_thread/core_interface/monothread";
-import MultiThreadContentInitializer from "../../../main_thread/init/multi_thread_content_initializer";
+import MediaSourceContentInitializer from "../../../main_thread/init/media_source_content_initializer";
 import metaplaylist from "../../../transports/metaplaylist";
 import addLocalManifestFeature from "../metaplaylist";
 
@@ -13,15 +13,15 @@ describe("Features list - METAPLAYLIST", () => {
     expect(featureObject).toEqual({
       transports: { metaplaylist },
       monothread: {
-        init: MultiThreadContentInitializer,
+        init: MediaSourceContentInitializer,
         coreInterface: MonoThreadCoreInterface,
-        workerMain: initializeWorkerMain,
+        initializeCoreEntry,
       },
     });
     expect(featureObject.transports.metaplaylist).toBe(metaplaylist);
     expect(featureObject.monothread).not.toBe(null);
-    expect(featureObject.monothread?.init).toBe(MultiThreadContentInitializer);
+    expect(featureObject.monothread?.init).toBe(MediaSourceContentInitializer);
     expect(featureObject.monothread?.coreInterface).toBe(MonoThreadCoreInterface);
-    expect(featureObject.monothread?.workerMain).toBe(initializeWorkerMain);
+    expect(featureObject.monothread?.initializeCoreEntry).toBe(initializeCoreEntry);
   });
 });
