+++ src/experimental/features/__tests__/multi_thread.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 import type { IFeaturesObject } from "../../../features/types";
 import { WorkerCoreInterface } from "../../../main_thread/core_interface/multithread";
-import MultiThreadContentInitializer from "../../../main_thread/init/multi_thread_content_initializer";
+import MediaSourceContentInitializer from "../../../main_thread/init/media_source_content_initializer";
 import addMultiThreadFeature from "../multi_thread";
 
 describe("Features list - EME", () => {
@@ -10,13 +10,13 @@ describe("Features list - EME", () => {
     addMultiThreadFeature(featureObject);
     expect(featureObject).toEqual({
       multithread: {
-        init: MultiThreadContentInitializer,
+        init: MediaSourceContentInitializer,
         coreInterface: WorkerCoreInterface,
       },
     });
     expect(featureObject.multithread).not.toBe(null);
     expect(featureObject.multithread).not.toBe(undefined);
-    expect(featureObject.multithread?.init).toBe(MultiThreadContentInitializer);
+    expect(featureObject.multithread?.init).toBe(MediaSourceContentInitializer);
     expect(featureObject.multithread?.coreInterface).toBe(WorkerCoreInterface);
   });
 });
