+++ src/features/list/__tests__/dash.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
-import initializeWorkerMain from "../../../core/main/worker";
+import initializeCoreEntry from "../../../core/entry";
 import { MonoThreadCoreInterface } from "../../../main_thread/core_interface/monothread";
-import MultiThreadContentInitializer from "../../../main_thread/init/multi_thread_content_initializer";
+import MediaSourceContentInitializer from "../../../main_thread/init/media_source_content_initializer";
 import dashJsParser from "../../../parsers/manifest/dash/js-parser";
 import DASHFeature from "../../../transports/dash";
 import type { IFeaturesObject } from "../../types";
@@ -19,15 +19,15 @@ describe("Features list - DASH", () => {
       transports: { dash: DASHFeature },
       dashParsers: { js: dashJsParser, wasm: null },
       monothread: {
-        init: MultiThreadContentInitializer,
+        init: MediaSourceContentInitializer,
         coreInterface: MonoThreadCoreInterface,
-        workerMain: initializeWorkerMain,
+        initializeCoreEntry,
       },
     });
     expect(featureObject.transports.dash).toBe(DASHFeature);
     expect(featureObject.monothread).not.toBe(null);
-    expect(featureObject.monothread?.init).toBe(MultiThreadContentInitializer);
+    expect(featureObject.monothread?.init).toBe(MediaSourceContentInitializer);
     expect(featureObject.monothread?.coreInterface).toBe(MonoThreadCoreInterface);
-    expect(featureObject.monothread?.workerMain).toBe(initializeWorkerMain);
+    expect(featureObject.monothread?.initializeCoreEntry).toBe(initializeCoreEntry);
   });
 });
