+++ src/features/list/dash_wasm.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-import initializeWorkerMain from "../../core/main/worker";
+import initializeCoreEntry from "../../core/entry";
 import type { IFeaturesObject } from "../../features/types";
 import { MonoThreadCoreInterface } from "../../main_thread/core_interface/monothread";
-import MultiThreadContentInitializer from "../../main_thread/init/multi_thread_content_initializer";
+import MediaSourceContentInitializer from "../../main_thread/init/media_source_content_initializer";
 import type { IDashWasmParserOptions } from "../../parsers/manifest/dash/wasm-parser";
 import DashWasmParser from "../../parsers/manifest/dash/wasm-parser";
 import dash from "../../transports/dash";
@@ -30,9 +30,9 @@ const dashWasmFeature = {
     }
     features.dashParsers.wasm = dashWasmParser;
     features.monothread = {
-      init: MultiThreadContentInitializer,
+      init: MediaSourceContentInitializer,
       coreInterface: MonoThreadCoreInterface,
-      workerMain: initializeWorkerMain,
+      initializeCoreEntry,
     };
   },
