+++ src/features/list/smooth.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import initializeWorkerMain from "../../core/main/worker";
+import initializeCoreEntry from "../../core/entry";
 import { MonoThreadCoreInterface } from "../../main_thread/core_interface/monothread";
-import MultiThreadContentInitializer from "../../main_thread/init/multi_thread_content_initializer";
+import MediaSourceContentInitializer from "../../main_thread/init/media_source_content_initializer";
 import smooth from "../../transports/smooth";
 import type { IFeaturesObject } from "../types";
 
@@ -29,9 +29,9 @@ function addSmoothFeature(features: IFeaturesObject): void {
     features.transports.smooth = smooth;
   }
   features.monothread = {
-    init: MultiThreadContentInitializer,
+    init: MediaSourceContentInitializer,
     coreInterface: MonoThreadCoreInterface,
-    workerMain: initializeWorkerMain,
+    initializeCoreEntry,
   };
 }
