+++ src/main_thread/api/public_api.ts
@@ -38,7 +38,9 @@ import type {
   IInbandEvent,
   IABRThrottlers,
   IBufferType,
+  ICoreMessage,
 } from "../../core/types";
+import { CoreMessageType } from "../../core/types";
 import type { IDefaultConfig } from "../../default_config";
 import type { IErrorCode, IErrorType } from "../../errors";
 import { ErrorCodes, ErrorTypes, formatError, MediaError } from "../../errors";
@@ -64,8 +66,6 @@ import {
   toVideoRepresentation,
   toAudioRepresentation,
 } from "../../manifest";
-import type { IWorkerMessage } from "../../multithread_types";
-import { MainThreadMessageType, WorkerMessageType } from "../../multithread_types";
 import type { IPlaybackObservation } from "../../playback_observer";
 import MediaElementPlaybackObserver from "../../playback_observer/media_element_playback_observer";
 import type {
@@ -132,6 +132,7 @@ import type { ContentInitializer } from "../init";
 import renderThumbnail from "../render_thumbnail";
 import type { IMediaElementTracksStore, ITSPeriodObject } from "../tracks_store";
 import TracksStore from "../tracks_store";
+import { MainThreadMessageType } from "../types";
 import type { IParsedLoadVideoOptions, IParsedStartAtOption } from "./option_utils";
 import {
   checkReloadOptions,
@@ -571,8 +572,8 @@ class Player extends EventEmitter<IPublicAPIEvent> {
         );
       };
       const handleInitMessages = (msg: MessageEvent) => {
-        const msgData = msg.data as unknown as IWorkerMessage;
-        if (msgData.type === WorkerMessageType.InitError) {
+        const msgData = msg.data as unknown as ICoreMessage;
+        if (msgData.type === CoreMessageType.InitError) {
           log.warn("API", "Processing InitError worker message: detaching worker");
           if (this._priv_worker !== null) {
             this._priv_worker.removeEventListener("message", handleInitMessages);
@@ -585,7 +586,7 @@ class Player extends EventEmitter<IPublicAPIEvent> {
               "Worker parser initialization failed: " + msgData.value.errorMessage,
             ),
           );
-        } else if (msgData.type === WorkerMessageType.InitSuccess) {
+        } else if (msgData.type === CoreMessageType.InitSuccess) {
           log.info("API", "InitSuccess received from worker.");
           if (this._priv_worker !== null) {
             this._priv_worker.removeEventListener("message", handleInitMessages);
@@ -1065,7 +1066,7 @@ class Player extends EventEmitter<IPublicAPIEvent> {
         log.info("API", "Initializing MediaSource mode in the main thread");
         const coreInterface = new features.monothread.coreInterface();
         const coreInterfaceCallbacks = coreInterface.getCallbacks();
-        features.monothread.workerMain(
+        features.monothread.initializeCoreEntry(
           coreInterfaceCallbacks.setCoreMessageReceiver,
           coreInterfaceCallbacks.sendCoreMessage,
         );
