+++ src/main_thread/init/media_source_content_initializer.ts
@@ -7,7 +7,12 @@ import type {
   IAdaptiveRepresentationSelectorArguments,
   IAdaptationChoice,
   IResolutionInfo,
+  ICreateMediaSourceCoreMessage,
+  ISentError,
+  ICoreMessage,
+  ISentLogValue,
 } from "../../core/types";
+import { CoreMessageType } from "../../core/types";
 import {
   EncryptedMediaError,
   MediaError,
@@ -25,17 +30,9 @@ import {
 } from "../../manifest";
 import MainMediaSourceInterface from "../../mse/main_media_source_interface";
 import type {
-  ICreateMediaSourceWorkerMessage,
-  ISentError,
-  ISentLogValue,
-  IWorkerMessage,
-} from "../../multithread_types";
-import { MainThreadMessageType, WorkerMessageType } from "../../multithread_types";
-import type {
   IReadOnlyPlaybackObserver,
   IMediaElementPlaybackObserver,
 } from "../../playback_observer";
-import type { IWorkerPlaybackObservation } from "../../playback_observer/worker_playback_observer";
 import type {
   ICmcdOptions,
   IInitialManifest,
@@ -60,8 +57,10 @@ import type { IContentProtection } from "../decrypt";
 import type IContentDecryptor from "../decrypt";
 import { ContentDecryptorState, getKeySystemConfiguration } from "../decrypt";
 import type { ITextDisplayer } from "../text_displayer";
+import { MainThreadMessageType } from "../types";
 import type { ITextDisplayerOptions } from "./types";
 import { ContentInitializer } from "./types";
+import type { ICorePlaybackObservation } from "./utils/create_core_playback_observer";
 import createCorePlaybackObserver from "./utils/create_core_playback_observer";
 import {
   resetMediaElement,
@@ -79,15 +78,15 @@ import { updateManifestCodecSupport } from "./utils/update_manifest_codec_suppor
 const generateContentId = idGenerator();
 
 /**
- * @class MultiThreadContentInitializer
+ * @class MediaSourceContentInitializer
  */
-export default class MultiThreadContentInitializer extends ContentInitializer {
-  /** Constructor settings associated to this `MultiThreadContentInitializer`. */
+export default class MediaSourceContentInitializer extends ContentInitializer {
+  /** Constructor settings associated to this `MediaSourceContentInitializer`. */
   private _settings: IInitializeArguments;
 
   /**
    * The Core may be sending messages as soon as we're preparing the content but
-   * the `MultiThreadContentInitializer` is only able to handle all of them only
+   * the `MediaSourceContentInitializer` is only able to handle all of them only
    * once `start`ed.
    *
    * As such `_queuedCoreMessages` is set to an Array  when `prepare` has been
@@ -97,17 +96,17 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
    * It is set to `null` when there's no need to rely on that queue (either not
    * yet `prepare`d or already `start`ed).
    */
-  private _queuedCoreMessages: IWorkerMessage[] | null;
+  private _queuedCoreMessages: ICoreMessage[] | null;
 
   /**
    * Information relative to the current loaded content.
    *
    * `null` when no content is prepared yet.
    */
-  private _currentContentInfo: IMultiThreadContentInitializerContentInfos | null;
+  private _currentContentInfo: IMediaSourceContentInitializerContentInfos | null;
   /**
    * `TaskCanceller` allowing to abort everything that the
-   * `MultiThreadContentInitializer` is doing.
+   * `MediaSourceContentInitializer` is doing.
    */
   private _initCanceller: TaskCanceller;
   /**
@@ -133,7 +132,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
       }
     >;
     /**
-     * Stores the resolvers and the current messageId that is sent to the core to
+     * Stores the resolvers and the current messageId that is sent to the web worker to
      * receive image thumbnails.
      */
     pendingThumbnailFetching: Map<
@@ -146,7 +145,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
   };
 
   /**
-   * Create a new `MultiThreadContentInitializer`, associated to the given
+   * Create a new `MediaSourceContentInitializer`, associated to the given
    * settings.
    * @param {Object} settings
    */
@@ -228,10 +227,10 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
     }
     this._queuedCoreMessages = [];
     log.debug("Init", "addEventListener prepare buffering core messages");
-    const onmessage = (msgData: IWorkerMessage): void => {
+    const onmessage = (msgData: ICoreMessage): void => {
       const type = msgData.type;
       switch (type) {
-        case WorkerMessageType.LogMessage: {
+        case CoreMessageType.LogMessage: {
           const formatted: IAcceptedLogValue[] = msgData.value.logs.map((l) => {
             switch (typeof l) {
               case "string":
@@ -512,9 +511,9 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
       );
     };
 
-    const onmessage = (msgData: IWorkerMessage) => {
+    const onmessage = (msgData: ICoreMessage) => {
       switch (msgData.type) {
-        case WorkerMessageType.AttachMediaSource: {
+        case CoreMessageType.AttachMediaSource: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -556,21 +555,21 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.Warning:
+        case CoreMessageType.Warning:
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
           this.trigger("warning", formatCoreError(msgData.value));
           break;
 
-        case WorkerMessageType.Error:
+        case CoreMessageType.Error:
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
           this._onFatalError(formatCoreError(msgData.value));
           break;
 
-        case WorkerMessageType.CreateMediaSource:
+        case CoreMessageType.CreateMediaSource:
           this._onCreateMediaSourceMessage(
             msgData,
             mediaElement,
@@ -579,7 +578,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           );
           break;
 
-        case WorkerMessageType.AddSourceBuffer:
+        case CoreMessageType.AddSourceBuffer:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -596,7 +595,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.SourceBufferAppend:
+        case CoreMessageType.SourceBufferAppend:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -639,7 +638,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.SourceBufferRemove:
+        case CoreMessageType.SourceBufferRemove:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -682,7 +681,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.AbortSourceBuffer:
+        case CoreMessageType.AbortSourceBuffer:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -703,7 +702,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.UpdateMediaSourceDuration:
+        case CoreMessageType.UpdateMediaSourceDuration:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -720,7 +719,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.InterruptMediaSourceDurationUpdate:
+        case CoreMessageType.InterruptMediaSourceDurationUpdate:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -737,7 +736,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.EndOfStream:
+        case CoreMessageType.EndOfStream:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -751,7 +750,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.InterruptEndOfStream:
+        case CoreMessageType.InterruptEndOfStream:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -765,7 +764,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.DisposeMediaSource:
+        case CoreMessageType.DisposeMediaSource:
           {
             if (
               this._currentContentInfo?.mediaSourceInfo?.type !== "main" ||
@@ -779,7 +778,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.NeedsBufferFlush: {
+        case CoreMessageType.NeedsBufferFlush: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -804,7 +803,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.ActivePeriodChanged: {
+        case CoreMessageType.ActivePeriodChanged: {
           if (
             this._currentContentInfo?.contentId !== msgData.contentId ||
             this._currentContentInfo.manifest === null
@@ -821,7 +820,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.AdaptationChanged: {
+        case CoreMessageType.AdaptationChanged: {
           if (
             this._currentContentInfo?.contentId !== msgData.contentId ||
             this._currentContentInfo.manifest === null
@@ -858,7 +857,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.RepresentationChanged: {
+        case CoreMessageType.RepresentationChanged: {
           if (
             this._currentContentInfo?.contentId !== msgData.contentId ||
             this._currentContentInfo.manifest === null
@@ -902,14 +901,14 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.EncryptionDataEncountered:
+        case CoreMessageType.EncryptionDataEncountered:
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
           lastContentProtection.setValue(msgData.value);
           break;
 
-        case WorkerMessageType.ManifestReady: {
+        case CoreMessageType.ManifestReady: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -920,7 +919,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.ManifestUpdate: {
+        case CoreMessageType.ManifestUpdate: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -942,14 +941,14 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.UpdatePlaybackRate:
+        case CoreMessageType.UpdatePlaybackRate:
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
           playbackObserver.setPlaybackRate(msgData.value);
           break;
 
-        case WorkerMessageType.BitrateEstimateChange:
+        case CoreMessageType.BitrateEstimateChange:
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -959,14 +958,14 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           });
           break;
 
-        case WorkerMessageType.InbandEvent:
+        case CoreMessageType.InbandEvent:
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
           this.trigger("inbandEvents", msgData.value);
           break;
 
-        case WorkerMessageType.LockedStream: {
+        case CoreMessageType.LockedStream: {
           if (
             this._currentContentInfo?.contentId !== msgData.contentId ||
             this._currentContentInfo.manifest === null
@@ -987,7 +986,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.PeriodStreamReady: {
+        case CoreMessageType.PeriodStreamReady: {
           if (
             this._currentContentInfo?.contentId !== msgData.contentId ||
             this._currentContentInfo.manifest === null
@@ -1058,7 +1057,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.PeriodStreamCleared: {
+        case CoreMessageType.PeriodStreamCleared: {
           if (
             this._currentContentInfo?.contentId !== msgData.contentId ||
             this._currentContentInfo.manifest === null
@@ -1072,7 +1071,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.DiscontinuityUpdate: {
+        case CoreMessageType.DiscontinuityUpdate: {
           if (
             this._currentContentInfo?.contentId !== msgData.contentId ||
             this._currentContentInfo.manifest === null
@@ -1098,7 +1097,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.PushTextData: {
+        case CoreMessageType.PushTextData: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -1124,7 +1123,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.RemoveTextData: {
+        case CoreMessageType.RemoveTextData: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -1156,7 +1155,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.ResetTextDisplayer: {
+        case CoreMessageType.ResetTextDisplayer: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -1171,7 +1170,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.StopTextDisplayer: {
+        case CoreMessageType.StopTextDisplayer: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -1186,7 +1185,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.ReloadingMediaSource:
+        case CoreMessageType.ReloadingMediaSource:
           {
             if (
               this._currentContentInfo === null ||
@@ -1210,7 +1209,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.NeedsDecipherabilityFlush:
+        case CoreMessageType.NeedsDecipherabilityFlush:
           {
             if (this._currentContentInfo?.contentId !== msgData.contentId) {
               return;
@@ -1235,7 +1234,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           }
           break;
 
-        case WorkerMessageType.SegmentSinkStoreUpdate: {
+        case CoreMessageType.SegmentSinkStoreUpdate: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -1250,15 +1249,15 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
           break;
         }
 
-        case WorkerMessageType.InitSuccess:
-        case WorkerMessageType.InitError:
+        case CoreMessageType.InitSuccess:
+        case CoreMessageType.InitError:
           // Should already be handled by the API
           break;
 
-        case WorkerMessageType.LogMessage:
+        case CoreMessageType.LogMessage:
           // Already handled by prepare's handler
           break;
-        case WorkerMessageType.ThumbnailDataResponse: {
+        case CoreMessageType.ThumbnailDataResponse: {
           if (this._currentContentInfo?.contentId !== msgData.contentId) {
             return;
           }
@@ -1622,7 +1621,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
       playbackObserver: IMediaElementPlaybackObserver;
     },
     cancelSignal: CancellationSignal,
-  ): IReadOnlyPlaybackObserver<IWorkerPlaybackObservation> | null {
+  ): IReadOnlyPlaybackObserver<ICorePlaybackObservation> | null {
     if (cancelSignal.isCancelled()) {
       return null;
     }
@@ -1917,7 +1916,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
    * @param {Object} coreInterface - The interface to the core.
    */
   private _onCreateMediaSourceMessage(
-    msg: ICreateMediaSourceWorkerMessage,
+    msg: ICreateMediaSourceCoreMessage,
     mediaElement: IMediaElement,
     mediaSourceStatus: SharedReference<MediaSourceInitializationStatus>,
     coreInterface: CoreInterface,
@@ -2003,7 +2002,7 @@ export default class MultiThreadContentInitializer extends ContentInitializer {
   }
 }
 
-export interface IMultiThreadContentInitializerContentInfos {
+export interface IMediaSourceContentInitializerContentInfos {
   /**
    * "contentId", which is the identifier for the currently loaded content.
    * Allows to ensure that the Core is referencing the current content, not
@@ -2015,6 +2014,7 @@ export interface IMultiThreadContentInitializerContentInfos {
    * `null` if not fetched / parsed yet.
    */
   manifest: IManifestMetadata | null;
+
   /**
    * Current MediaSource linked to the content.
    *
@@ -2079,7 +2079,7 @@ export interface IMultiThreadContentInitializerContentInfos {
 /** Arguments to give to the `InitializeOnMediaSource` function. */
 export interface IInitializeArguments {
   /**
-   * The `MultiThreadContentInitializer` will interact with the RxPlayer's core
+   * The `MediaSourceContentInitializer` will interact with the RxPlayer's core
    * logic (the one loading media data) by exchanging messages through an
    * interface called the `CoreInterface`.
    *
@@ -2244,7 +2244,7 @@ function formatCoreError(sentError: ISentError): IPlayerError {
         tracks: sentError.tracks,
       });
     case "EncryptedMediaError":
-      // We assume that everything have already been checked core-side here
+      // We assume that everything have already been checked Worker-side here
       // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
       return new EncryptedMediaError(sentError.code, sentError.reason, {
         keyStatuses: sentError.keyStatuses,
@@ -2337,6 +2337,7 @@ function formatSentLogObject(arg: ISentLogValue): IAcceptedLogValue {
   if (typeof arg !== "object") {
     return arg;
   }
+
   if (arg?.isSerializedError === true) {
     return formatCoreError(arg as ISentError);
   }
