+++ src/main_thread/init/utils/create_core_playback_observer.ts
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-import getBufferedDataPerMediaBuffer from "../../../core/main/common/get_buffered_data_per_media_buffer";
+// TODO: move that file somewhere else
+import getBufferedDataPerMediaBuffer from "../../../core/entry/get_buffered_data_per_media_buffer";
 import type { IPausedPlaybackObservation } from "../../../core/types";
 import type { IManifestMetadata } from "../../../manifest";
 import { getMaximumSafePosition } from "../../../manifest";
