+++ src/worker_entry_point.ts
@@ -3,12 +3,12 @@
  * on when running in a multithread mode.
  */
 
-import initializeWorkerMain from "./core/main/worker";
+import initializeCoreEntry from "./core/entry";
+import type { ICoreMessage } from "./core/types";
+import { CoreMessageType } from "./core/types";
 import log from "./experimental/tools/mediaCapabilitiesProber/log";
 import features from "./features";
 import Manifest from "./manifest/classes";
-import type { IWorkerMessage } from "./multithread_types";
-import { WorkerMessageType } from "./multithread_types";
 import DashJsParser from "./parsers/manifest/dash/js-parser";
 import DashWasmParser from "./parsers/manifest/dash/wasm-parser";
 import createDashPipelines from "./transports/dash";
@@ -24,7 +24,7 @@ features.transports.dash = createDashPipelines;
 globalScope.onmessageerror = (_msg: MessageEvent) => {
   log.error("Core", "Error when receiving message from main thread.");
 };
-initializeWorkerMain((handler) => {
+initializeCoreEntry((handler) => {
   onmessage = handler;
 }, sendMessage);
 
@@ -34,7 +34,7 @@ initializeWorkerMain((handler) => {
  * @param {Object} msg
  * @param {Array.<Object>} [transferables]
  */
-function sendMessage(msg: IWorkerMessage, transferables?: Transferable[]): void {
+function sendMessage(msg: ICoreMessage, transferables?: Transferable[]): void {
   updateMessageFormat(msg);
 
   log.debug("M<--C", "Sending message from worker", { name: msg.type });
@@ -42,7 +42,7 @@ function sendMessage(msg: IWorkerMessage, transferables?: Transferable[]): void 
     postMessage(msg);
   } else {
     // TypeScript made a mistake here, and 2busy2fix
-    (postMessage as (msg: IWorkerMessage, transferables: Transferable[]) => void)(
+    (postMessage as (msg: ICoreMessage, transferables: Transferable[]) => void)(
       msg,
       transferables,
     );
@@ -56,14 +56,14 @@ function sendMessage(msg: IWorkerMessage, transferables?: Transferable[]): void 
  * If necessary, mutations are done in place.
  * @param {Object} msg
  */
-function updateMessageFormat(msg: IWorkerMessage): void {
+function updateMessageFormat(msg: ICoreMessage): void {
   if (
-    msg.type === WorkerMessageType.ManifestReady ||
-    msg.type === WorkerMessageType.ManifestUpdate
+    msg.type === CoreMessageType.ManifestReady ||
+    msg.type === CoreMessageType.ManifestUpdate
   ) {
     if (msg.value.manifest instanceof Manifest) {
       msg.value.manifest = msg.value.manifest.getMetadataSnapshot();
-      if (msg.type === WorkerMessageType.ManifestUpdate) {
+      if (msg.type === CoreMessageType.ManifestUpdate) {
         // Remove `periods` key to reduce cost of an unnecessary manifest
         // clone.
         msg.value.manifest.periods = [];
